Exposing A, the atom_marker in the OutputSystem
We can now also print stuff for beam entities etc. - just add the marker as the third argument.

I also changed that in the example files that only print out text because I did not find a way to do it for SerdeJson only.

Furthermore, an example of a non-atom writeout has been added to 2d_plus_mot_from_oven.rs